Add BIND_INTERFACE env var (#73)
Allows setting the `bind-interface` config option in fr24feed.ini
to allow access from non-private IP addresses. Can be set to
`0.0.0.0` to allow access from any IP address (warning, may be
insecure). If env var is not set (default), the config option
will not be added to the config file, which is a secure default.

Fixes #58